challenge/bank-apis-main/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
motor==3.3.2
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.0
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
